+++ R/draws-index.R
@@ -1,27 +1,18 @@
 #' Index `draws` objects
 #'
-#' Index variables, iterations, chains, and draws.
+#' Index iterations, chains, and draws of [`draws`] objects.
 #'
 #' @name draws-index
 #' @template args-methods-x
-#' @template args-methods-dots
-#' @param value (character vector) For `variables(x) <- value`, the new variable
-#'   names to use.
 #'
 #' @details
-#' The methods `variables()`, `iteration_ids()`, `chain_ids()`, and `draw_ids()` return
-#' vectors of all variables, iterations, chains, and draws, respectively. In
-#' contrast, the methods `nvariables()`, `niterations()`, `nchains()`, and
+#' The methods `iteration_ids()`, `chain_ids()`, and `draw_ids()` return
+#' vectors of all iterations, chains, and draws, respectively. In
+#' contrast, the methods `niterations()`, `nchains()`, and
 #' `ndraws()` return the number of variables, iterations, chains, and draws,
 #' respectively.
 #'
-#' `variables(x) <- value` allows you to modify the vector of variable names,
-#' similar to how `names(x) <- value` works for vectors and lists. For renaming
-#' specific variables, [set_variables()] works equivalently, but is more intuitive when using the pipe operator. [rename_variables()] may offer a more convenient approach.
-#'
 #' @return
-#'
-#' For `variables()`, a character vector.
 #'
 #' For `iteration_ids()`, `chain_ids()`, and `draw_ids()`, an integer vector.
 #'
@@ -31,10 +22,6 @@
 #'
 #' @examples
 #' x <- example_draws()
-#'
-#' variables(x)
-#' nvariables(x)
-#' variables(x) <- letters[1:nvariables(x)]
 #'
 #' iteration_ids(x)
 #' niterations(x)
@@ -46,133 +33,6 @@
 #' ndraws(x)
 #'
 NULL
-
-#' @rdname draws-index
-#' @export
-variables <- function(x, ...) {
-  UseMethod("variables")
-}
-
-#' @export
-variables.NULL <- function(x, ...) {
-  NULL
-}
-
-#' @export
-variables.draws_matrix <- function(x, reserved = FALSE, ...) {
-  remove_reserved_variable_names(colnames(x), reserved)
-}
-
-#' @export
-variables.draws_array <- function(x, reserved = FALSE, ...) {
-  remove_reserved_variable_names(dimnames(x)[[3L]], reserved)
-}
-
-#' @export
-variables.draws_df <- function(x, reserved = FALSE, ...) {
-  # reserved_df_variables are special data.frame columns
-  # which should never be included as variables
-  out <- names(x)[!names(x) %in% reserved_df_variables()]
-  remove_reserved_variable_names(out, reserved)
-}
-
-#' @export
-variables.draws_list <- function(x, reserved = FALSE, ...) {
-  if (!length(x)) {
-    return(character(0))
-  }
-  remove_reserved_variable_names(names(x[[1]]), reserved)
-}
-
-#' @export
-variables.draws_rvars <- function(x, reserved = FALSE, ...) {
-  remove_reserved_variable_names(names(x), reserved)
-}
-
-# remove reserved variable names
-remove_reserved_variable_names <- function(variables, reserved) {
-  reserved <- as_one_logical(reserved)
-  if (!reserved && length(variables)) {
-    # can't use setdiff() here as in the edge case where someone
-    # manually creates duplicate columns it will give incorrect results
-    variables <- variables[!variables %in% reserved_variables()]
-  }
-  variables
-}
-
-#' @rdname draws-index
-#' @export
-`variables<-` <- function(x, value) {
-  UseMethod("variables<-")
-}
-
-#' @export
-`variables<-.draws_matrix` <- function(x, value) {
-  check_new_variables(value)
-  colnames(x) <- value
-  x
-}
-
-#' @export
-`variables<-.draws_array` <- function(x, value) {
-  check_new_variables(value)
-  dimnames(x)[[3L]] <- value
-  x
-}
-
-#' @export
-`variables<-.draws_df` <- function(x, value) {
-  check_new_variables(value)
-  names(x)[!names(x) %in% reserved_df_variables()] <- value
-  x
-}
-
-#' @export
-`variables<-.draws_list` <- function(x, value) {
-  check_new_variables(value)
-  for (i in seq_along(x)) {
-    names(x[[i]]) <- value
-  }
-  x
-}
-
-#' @export
-`variables<-.draws_rvars` <- function(x, value) {
-  check_new_variables(value)
-  names(x) <- value
-  x
-}
-
-#' Set variable names in `draws` objects
-#'
-#' Set variable names for all variables in a [`draws`] object. Useful
-#' when using pipe operators.
-#'
-#' @param x (draws) A [`draws`] object.
-#' @param variables (character) new variable names.
-#' @template args-methods-dots
-#'
-#' @return Returns a [`draws`] object of the same format as `x`, with
-#'   variables named as specified.
-#'
-#' @seealso [`variables`]
-#' @examples
-#' x <- as_draws(matrix(rnorm(100), ncol = 2))
-#' variables(x)
-#'
-#' x <- set_variables(x, c("theta[1]", "theta[2]"))
-#' variables(x)
-#'
-#' # this is equivalent to
-#' variables(x) <- c("theta[1]", "theta[2]")
-#' variables(x)
-#'
-#' @export
-set_variables <- function(x, variables, ...) {
-  variables(x) <- variables
-  return(x)
-}
-
 
 #' @rdname draws-index
 #' @export
@@ -320,21 +180,6 @@ draw_ids.rvar <- function(x) {
 
 #' @rdname draws-index
 #' @export
-nvariables <- function(x, ...) {
-  UseMethod("nvariables")
-}
-#' @export
-nvariables.NULL <- function(x, ...) {
-  0
-}
-
-#' @export
-nvariables.draws <- function(x, ...) {
-  length(variables(x, ...))
-}
-
-#' @rdname draws-index
-#' @export
 niterations <- function(x) {
   UseMethod("niterations")
 }
@@ -471,106 +316,6 @@ ndraws.rvar <- function(x) {
 
 # internal ----------------------------------------------------------------
 
-# check validity of existing variable names: e.g., that
-# all `variables` exist in `x` and that no `variables`are reserved words
-# Additionally, this returns the cannonical name, so e.g. "theta" will get
-# converted to c("theta[1]", "theta[2]", ...) if those variables exist.
-# @param regex should 'variables' be treated as regular expressions?
-# @param scalar_only should only scalar variables be matched?
-check_existing_variables <- function(variables, x, regex = FALSE,
-                                     scalar_only = FALSE, exclude = FALSE) {
-  check_draws_object(x)
-  if (is.null(variables)) {
-    return(NULL)
-  }
-
-  regex <- as_one_logical(regex)
-  scalar_only <- as_one_logical(scalar_only)
-  exclude <- as_one_logical(exclude)
-  variables <- unique(as.character(variables))
-  all_variables <- variables(x, reserved = TRUE)
-  
-  if (regex) {
-    tmp <- named_list(variables)
-    for (i in seq_along(variables)) {
-      tmp[[i]] <- grep(variables[i], all_variables)
-    }
-    # regular expressions are not required to match anything
-    missing_variables <- NULL
-    variables <- as.character(all_variables[unique(unlist(tmp))])
-  } else if (!scalar_only) {
-    # need to find variables that are matched by either a scalar or vector
-    # variable in x and what the matching variable is, while keeping original
-    # order of input `variables`
-
-    # find scalar variables (1-to-1 match between all_variables and variables)
-    scalar_input_ixs <- match(all_variables, variables)
-    # find vector variable matches (match all_variables with the indexing stripped)
-    all_variables_base <- all_variables
-    # exclude already matched scalar variables
-    all_variables_base[!is.na(scalar_input_ixs)] <- NA_character_
-    all_variables_base <- gsub("\\[.*\\]$", "", all_variables_base, perl = TRUE)
-    vector_input_ixs <- match(all_variables_base, variables)
-    # compose the vector of indices of matched input variables
-    input_ixs <- c(scalar_input_ixs[!is.na(scalar_input_ixs)],
-                   vector_input_ixs[!is.na(vector_input_ixs)])
-    # compose the vector of indices of matched all_variables
-    all_var_ixs <- seq_along(all_variables)
-    all_var_matched_ixs <- c(all_var_ixs[!is.na(scalar_input_ixs)],
-                             all_var_ixs[!is.na(vector_input_ixs)])
-    # select missed input variables
-    missing_vars_mask <- rep_len(TRUE, length(variables))
-    missing_vars_mask[input_ixs] <- FALSE
-    missing_variables <- variables[missing_vars_mask]
-    # select matched all_variables maintaining the input variables order
-    variables <- all_variables[all_var_matched_ixs[order(input_ixs, all_var_matched_ixs)]]
-  } else {
-    missing_variables <- setdiff(variables, all_variables)
-  }
-  variables <- check_reserved_variables(variables)
-  if (length(missing_variables)) {
-    stop_no_call("The following variables are missing in the draws object: ",
-          comma(missing_variables))
-  }
-
-  # handle excluding variables for subset_draws
-  if (exclude) {
-    variables <- setdiff(all_variables, variables)
-  }
-  
-  invisible(variables)
-}
-
-# check validity of new variables: e.g., that there are
-# no duplicates in `variables` and that they do not use
-# reserved words
-check_new_variables <- function(variables) {
-  # use anyDuplicated() for the check since it is faster than any(duplicated(x)) and
-  # we shouldn't expect to take this branch often (since it is an error)
-  if (anyDuplicated(variables)) {
-    duplicates = unique(variables[duplicated(variables)])
-    stop_no_call(
-      "Duplicate variable names are not allowed in draws objects.\n",
-      "The following variable names are duplicates:\n",
-      comma(duplicates)
-    )
-  }
-  check_reserved_variables(variables)
-}
-
-# check variables do not make use of reserved words
-check_reserved_variables <- function(variables) {
-  assert_character(variables)
-  # for now only check reserved columns used in 'draws_df' objects
-  # other reserved variables such as '.log_weight' may be overwritten
-  # this has the advantage that power users can directly add such variables
-  used_reserved_variables <- intersect(reserved_df_variables(), variables)
-  if (length(used_reserved_variables)) {
-    stop_no_call("Variable names ", comma(used_reserved_variables), " are reserved.")
-  }
-  invisible(variables)
-}
-
 # check validity of iteration indices
 # @param unique should the returned IDs be unique?
 check_iteration_ids <- function(iteration_ids, x, unique = TRUE, exclude = FALSE) {
@@ -599,7 +344,7 @@ check_iteration_ids <- function(iteration_ids, x, unique = TRUE, exclude = FALSE
   if (exclude) {
     iteration_ids <- setdiff(iteration_ids(x), iteration_ids)
   }
-  
+
   invisible(iteration_ids)
 }
 
@@ -630,7 +375,7 @@ check_chain_ids <- function(chain_ids, x, unique = TRUE, exclude = FALSE) {
   if (exclude) {
     chain_ids <- setdiff(chain_ids(x), chain_ids)
   }
-  
+
   invisible(chain_ids)
 }
 
@@ -661,6 +406,6 @@ check_draw_ids <- function(draw_ids, x, unique = TRUE, exclude = FALSE) {
   if (exclude) {
     draw_ids <- setdiff(draw_ids(x), draw_ids)
   }
-  
+
   invisible(draw_ids)
 }